Better Crash Consistency
Updates to perform better crash consistency for data in flight.

Fixes to ensure consistency of hashtable and data after crash.

Data flow as follows
1. Data written to archive on write without sync option
2. Hash and ref count committed to memory hashtable
3. Archive Closed and become immutable
4. Archive is sync'd to filesystem
5. Archive Map is sync'd to filesystem
6. Archive and ref count is committed to persistent hashtable (rocksdb)
7. Archive is committed to cloud

On crash data that is not sync'd to local filesystem is ignored in
persistent hashtable because it has not been persisten but available for
read from the few files that were being written at the time of crash.
All data is still uploaded to cloud but is not referenced by hashtable.
This ensures corrupt data does not poison the hashtable but uses best
effort for files that were writing at time of crash.